+++ cpukit/score/include/rtems/score/schedulerpriorityimpl.h
@@ -210,13 +210,13 @@ RTEMS_INLINE_ROUTINE void _Scheduler_priority_Ready_queue_update(
   Chain_Control                  *ready_queues
 )
 {
-  ready_queue->current_priority = new_priority;
+  ready_queue->current_priority = (unsigned int) new_priority;
   ready_queue->ready_chain = &ready_queues[ new_priority ];
 
   _Priority_bit_map_Initialize_information(
     bit_map,
     &ready_queue->Priority_map,
-    new_priority
+    (unsigned int) new_priority
   );
 }
